backend/functions/requirements.txt
firebase_functions
firebase_admin
google_cloud_firestore